chore: refactor basic file structure
