+++ main.go
@@ -627,22 +627,3 @@ func leftPadTo32Bytes(i *big.Int) []byte {
 	copy(padded[32-len(bytes):], bytes)
 	return padded
 }
-
-func banner() string {
-	b := "\n"
-	b += fmt.Sprintf("%s%s                                     %s\n", ansiCodes["invertOn"], ansiCodes["bold"], ansiCodes["reset"])
-	b += fmt.Sprintf("%s%s     io.finnet Key Recovery Tool     %s\n", ansiCodes["invertOn"], ansiCodes["bold"], ansiCodes["reset"])
-	b += fmt.Sprintf("%s%s               v5.0.0                %s\n", ansiCodes["invertOn"], ansiCodes["bold"], ansiCodes["reset"])
-	b += fmt.Sprintf("%s%s                                     %s\n", ansiCodes["invertOn"], ansiCodes["bold"], ansiCodes["reset"])
-	b += "\n"
-	return b
-}
-
-func errorBox(err error) string {
-	b := "\n"
-	b += fmt.Sprintf("%s%s         %s\n", ansiCodes["darkRedBG"], ansiCodes["bold"], ansiCodes["reset"])
-	b += fmt.Sprintf("%s%s  Error  %s  %s.\n", ansiCodes["darkRedBG"], ansiCodes["bold"], ansiCodes["reset"], err)
-	b += fmt.Sprintf("%s%s         %s\n", ansiCodes["darkRedBG"], ansiCodes["bold"], ansiCodes["reset"])
-	b += "\n"
-	return b
-}
